fix(core): RemovalPolicy.SNAPSHOT can be added to resources that do not support it (#20668)
fixes #20653, which was reported for opensearch domains, but also fixes this issue for all other resources that allowed this removal policy but did not actually support it.

previously, the behavior was a successful synth and deploy, and cloudformation ignores the snapshot removal policy if it is not supported by the resource. because of this, i believe that i have to add this fix behind a feature flag to avoid breaking customers.

without the feature flag, the best we can do is add a warning to the stack notifying customers that the snapshot policy is useless.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*