feat(apigateway): Add LambdaIntegrationOptions to LambdaRestApi (#17065)
This adds an `integrationOptions` field to the `LambdaRestApiProps` and ensures that `proxy` cannot be set differently from the base `proxy` flag (but discourages setting it at all). This will allow setting the `cacheKeyParameters`, `allowTestInvoke`, and `vpcLink` configuration fields without having to fall back to using a regular `RestApi`.

Some of this is inspired by #12004 and its review comments, which was closed earlier this year due to inactivity.

Resolves: #3269

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*