fix(pipelines): 'ConfirmPermissionsBroadening' uses wrong node version
The CodeBuild Project that is used when the `ConfirmPermissionsBroadening`
feature is enabled does not have a CodeBuild image specified. This
makes it use the `standard:2.0` image by default, which is AL2-based
and comes with Node 12. CDK tooling now requires Node 14 to run.

Make it use the `standard:5.0` image like all other CodeBuild projects,
which contains Node 14.