chore(cfnspec): new libraries do not conform to pkglint (#20900)
When we updated the minimum Node version, we did not update it
in the template that is used to stamp out new construct libraries
when new CFN namespaces are added to the spec.

Meaning every cfnspec update now fails because the new packages
fail `pkglint` validation.

Update the Node version in the template.

Closes #20856.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*