feat(aws-s3): create default bucket policy when required (under feature flag) (#20765)
Created a new feature flag `@aws-cdk/aws-s3:createDefaultLoggingPolicy`

There are certain types of S3 Buckets that AWS will automatically create a bucket policy for you if you do not create one. For example, if you create an S3 Bucket to be used as the destination for VPC Flow Logs and you  do not create a Bucket Policy, AWS will automatically create a bucket policy for you. The full list of resources can be found [here](https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/AWS-logs-and-resource-policy.html#AWS-logs-infrastructure-S3)

- [X] Vpc Flow Logs
~- [ ] AWS Network Firewall logs~ (No L2 support yet)
~- [ ] AWS Global Accelerator flow logs~ ([not currently possible](https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/922)]
~- [ ] EC2 Spot Instance data feed~(no cloudformation support yet)
~- [ ] CloudFront access logs & streaming access logs~ (CloudFront uses bucket ACL _not_ bucket policy)
- [X] Network Load Balancer access logs (already done)
- [x] Amazon Managed Streaming for Apache Kafka broker logs

If we allow AWS to create these policies automatically, it prevents CDK from every managing that policy in the future. Since we know what the policy should be we should instead create the logging bucket with the required policy.

fixes #18816


----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*