fix(cli): CLI errors when run as a non-existent user
The CDK home directory was being created in the users home directory,
but this causes an error if for some reason the user does not have a
home directory. Since the cdk home directory is only being used to
create a cache directory, this PR updates the location from being the
users home directory to being the os tmp directory.

I don't think there
is a strong reason why this directory needs to live in the users home
directory, but if there is I could also change this to have some
fallback logic instead.

fix #7937