chore(cli): remove scary "you shouldn't see this" message
We thought people shouldn't see this message, but they occasionaly do, which leads
to alarmed bug reports.

It's *probably* okay to remove the `on('timeout')` event handler if we
have to have this one anyway, but I'm not a 100% confident so I'd rather
just make the message less objectionable and have 2 timers, than have 1
and not be sure what's going on.