chore(init-templates): remove v1 init templates (#21019)
A v1 app or library cannot be created from the v2 cli. This cleans up the v1 init template code.

This is stupidly long because of the amount of files deleted and moved so here's a rundown of the changes:
- v1 templates deleted (github is showing some of these as renamed because they were duplicates of the v2 files that did get moved)
- v2 templates removed from `v2` folder and moved directly under `init-templates`
- hook files imports updated (another place where it shows the change to v1 instead of v2 because the contents were the same) for templates and sample apps
- `init.ts`: updated file paths to flattened directory
- `init.test.ts`: 
    - updated tests to only run for v2 templates
    - added test for `fsharp` since it was missing
    - updated .NET tests to look in the contents of the `.sln` file for the project information. I noticed that this test sometimes still passed when the `init` actually failed.


Note: This is the first of two PRs to fix the bug for creating .NET projects with spaces in the file path (issue #18803). This PR was done separately so the bug fix didn't get lost in the 300 file changes due to the cleanup.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*