fix(init-templates): csharp and fsharp init fails when path contains space
When running  or  with a space in the path to the project, init will fail. This fix adds in handling for spaces and other special characters in the filepath for both windows systems and posix systems. Tests have been added for posix and manual testing was performed on a windows machine.

Closes issue #18803.