docs(logs): Clarify encrpytion default for Log Groups
The default encryption for CloudWatch Logs is not performed via a KMS
AWS-managed key which is what is usually implied by "master key".
Updating the documentation to clarify that this is server-side
encryption hopefully clarifies why "KMS key ID" shows as "-" in the
Console for created Log Groups and why the
`cloudwatch-log-group-encrypted` Config Rule shows as non-compliant.