fix(cli): revert "fix(cli): format of tags in cdk.json is not validated" (#21092)
Reverts aws/aws-cdk#21050

It breaks the CLI integ tests that get run on our internal pipelines. The error message is `Error: tags must be an array of { Tag: string, Value: string } objects`, which comes from the `validateTags` method in #21050. The test runs `cdk deploy --tag key=value` which I was able to test on my own account to verify that that is a valid way of adding tags to a stack.