fix(aws-eks): cap generated stack names at 128 characters (#20528)
When imported, Kubectl Providers result in a new nested stack being created. Nested stack names are created by generating a unique ID (which includes the entire construct id path). Kubectl provider ids are constructed the same way (they can't be specified manually, so the code uses a unique ID). This results in the final ID for the stack containing the id of the provider repeated twice. The makes the name hard to decipher, and more importantly can result in stack names being longer than 128 characters.

This caps all generated stack names at 128 characters. Ideally, we would be able to change way in which these stack names are generated; however, that would be a breaking change.

Fixes #20124.

----


*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*