Revert "fix(custom-resources): Custom resource provider framework not passing `ResponseURL` to user function" (#21109)
This PR did not actually fix the issue with custom resources, it just passed a `ResponseURL: '...'`, which isn't a usable URL. Leaving it as a sanitized string makes it appear as if the `ResponseURL` is available, but it actually isn't. Reverting this PR to not give the false impression that `ResponseURL` is usable.

Reverts aws/aws-cdk#21065