fix(custom-resources): Custom resource provider framework not passing `ResponseURL` to user function (aws#21065)
aws#20889 included a change that broke the custom resource framework by not including the necessary presigned URL. We attempted to fix this with aws#21065 but it didn't resolve the issue and aws#21109 reverted the attempted fix.

This changes explicitly includes the presigned URL, as well as adding tests to ensure the URL is passed to the downstream Lambda Function.

Closes aws#21058