docs(logs): clarify encryption default for log groups (#21060)
The default encryption for CloudWatch Logs is not performed via a KMS
AWS-managed key which is what is usually implied by "master key".
Updating the documentation to clarify that this is server-side
encryption hopefully clarifies why "KMS key ID" shows as "-" in the
Console for created Log Groups and why the
`cloudwatch-log-group-encrypted` Config Rule shows as non-compliant.

Based on https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/data-protection.html#encryption-rest

----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*