feat(backup): support RDS database cluster and serverless cluster (#17971)
Add methods `fromRdsDatabaseCluster` and `fromRdsServerlessCluster` to support RDS database cluster and serverless cluster in AWS Backup. In `BackupableResourcesCollector`, `CfnDBCluster` will be detected and added to the backup selection.

Missing unit tests for `fromRdsDatabaseInstance` were added as well.

Change behaviour of `CfnDBInstance` in `BackupableResourcesCollector`. Database instances will be added only if attribute `dbClusterIdentifier` is not set. In the current implementation, database instances of a database cluster will be added to the backup selection. In my point of view, it is not necessary to backup each database instance of a database cluster if the database cluster is selected. Please check if this change is correct and allowed.

Closes #16457.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*