feat(cli): allow diffing against a processed template (#19908)
In some cases, using an unprocessed template with `CfnInclude` fails,
most commonly because the unprocessed template can contain cycles between the resources.
When this happens, customers need to include the processed template instead,
but when that happens, the `diff` command returns the wrong results,
because it compares against the original, and thus unprocessed, template.

Add a new command-line switch to the `diff` command that allows comparing against the processed template,
which gives the correct results in this case.

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*