chore: normalize test names for integration tests
This PR is a follow up to #21093 which updated the logic to normalize
the name of the integration test that was stored in the `integ.json`
file. This PR updates the names in all of the `integ.json` files.

This will cut down on some of the noise/confusion in PRs.