fix(cli): large context causes E2BIG error during synthesis on Linux
Instead of passing the context in an environment variable, the CLI now writes the context to a temporary file and sets an environment variable only with the location. The app then uses that location to read from the file.

Also tested manually on a Linux machine.

Re-roll https://github.com/aws/aws-cdk/pull/21230
Fixes https://github.com/aws/aws-cdk/issues/19261