feat(cli): cannot pass objects and numbers as context arguments  (#20068)
All context arguments passed to the cli are parsed as strings.

```
cdk deploy --context @aws-cdk/core:newStyleStackSynthesis=false
cdk deploy --context @aws-cdk/core:newStyleStackSynthesis=0
```

The values are interpreted as strings so the above two values get
interpreted as 'false' and '0' which are truthy (and not what we want).

This PR adds a function to parse context arguments and try to covert values
into their correct type.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*