feat(core): use inode data to fingerprint large assets
When fingerprinting large assets, hashing the asset can take quite a
long time - over a second for a 300MB asset, for example. This can add
up, particularly when generating multiple stacks in a single build, or
when running test suites that bundle assets multiple times, and is not
avoidable by asset caching (since it's computing the cache key).

This change switches to using file inode, mtime, and size instead of
content to fingerprint files larger than 16MB (with a new context key
that can be used to configure this behavior).

This feature improved the runtime of one of our slowest tests by ~10%.

closes: #21297