chore(core): use inode data to cache fingerprints
When fingerprinting large assets, hashing the asset can take quite a
long time - over a second for a 300MB asset, for example. This can add
up, particularly when generating multiple stacks in a single build, or
when running test suites that bundle assets multiple times, and is not
avoidable by asset caching (since it's computing the cache key).

This change caches the result of digesting individual files based on the
inode, mtime, and size of the input file.

This feature improved the runtime of one of our slowest tests by ~10%.

closes: #21297