feat(cli): support hotswapping Lambda function's description and environment variables
This change allows Lambda function to be hotswap'ed when there's change
in the function's description and/or environment variables. These changes
are categorized as configuration changes and are updated by calling
`updateFunctionConfiguration`. Since the existing waiter
"UpdateFunctionCodeToFinish" is now used to wait for both code update
and configuration update, I renamed it to "UpdateFunctionPropertiesToFinish".