feat(custom-resource): allow AwsCustomResource to be placed in vpc
This will allow for users with a need for all Lambda functions to be
placed in a VPC to more easily adopt the use of `AwsCustomResource`.
This is implemented by accepting a `vpc` and `vpcSubnets` property and
passing those up to the created `SingletonFunction`. Then tests are
added to ensure that it builds correctly and that errors are thrown in
the expected circumstances (in the same situations that
`lambda.Function` does).