fix(aws-lambda): FunctionUrl incorrectly uses Alias ARNs
When configuring a FunctionUrl for a Lambda function Alias,
the underlying lambda function's ARN must be configured
as TargetFunctionArn, and the alias name as Qualifier. This
was previously not the case, as the Alias' ARN was used
with no qualifier, which accidentally succeeded provisionning
but did not necessarily produce the intended result.