feat(core): cache fingerprints of large assets (#21321)
When fingerprinting large assets, hashing the asset can take quite a
long time - over a second for a 300MB asset, for example. This can add
up, particularly when generating multiple stacks in a single build, or
when running test suites that bundle assets multiple times, and is not
avoidable by asset caching (since it's computing the cache key).

This change caches the result of digesting individual files based on the
inode, mtime, and size of the input file.

This feature improved the runtime of one of our slowest tests by ~10%.

closes: #21297

Note: No README entries were added, because this sub-subsystem was already not documented in the README.


----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)? N/A