feat(rds): add copyTagsToSnapshot to the construct props for ServerlessCluster and ServerlessClusterFromSnapshot (#21056)
In f7c628948e7f71df7a95cb00cdc2746e2e46dc03 copyTagsToSnapshot was added to DatabaseCluster and in 40a6ceb9983694a3645edd78167e93825a9049e9 it was added to DatabaseClusterFromSnapshot as well.

This does the same for ServerlessCluster and ServerlessClusterFromSnapshot.

The implementation and tests parallels the changes mentioned above.

fixes #20968

## Behaviour change

In parallel to copyTagsToSnapshot on DatabaseCluster and DatabaseClusterFromSnapshot, the default for copyTagsToSnapshot on ServerlessCluster and ServerlessClusterFromSnapshot is true.

Before this change, ServerlessCluster and ServerlessClusterFromSnapshot did not emit the CopyTagsToSnapshot property in the CloudFormation template. That resulted in the CopyTagsToSnapshot property being false.

My rationale for this behaviour change is to minimise the differences between DatabaseCluster[FromSnapshot] and ServerlessCluster[FromSnapshot], which are both realised as the same L1 construct.

Changing the CopyTagsToSnapshot property on an existing serverless RDS cluster introduces no interruption. Ref: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-rds-dbcluster.html#cfn-rds-dbcluster-copytagstosnapshot

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*