docs(dynamodb): the documentation is wrong around table encryption (#20938)
----
This pull request aims to resolve https://github.com/aws/aws-cdk/issues/16436. As mentioned in the issue, by default, no encryption is done, which is different from the documentation description. Now the documentation descriptions are updated around the table encryption as the default is for the Table to not be encrypted.

Closes #16436

----
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*