fix(appmesh): routes with weight 0 are assigned a weight of 1 (#21400)
----

### All Submissions:

fixes #21365

This PR fixes an unexpected behaivor when creating an appmesh.

### Expected Behavior
Routes with targets of weight 0 should not change to 1 automatically.

### Current Behavior (without fix)
Routes with targets of weight 0 are being changed to 1 automatically.

### Solution

`t.weight || 1` will result in `0 || 1` if the `weight` is set to `0`. This will result in a route that can never have a weight of 0, since `0` is a `faulty` value.

To fix this, I compare the weight with `undefined` & return `1` only if the weight is undefined.

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*